Refactor & start working on project module

+++ src/common/entities/common.entity.go
@@ -3,7 +3,7 @@ package entities
 import "time"
 
 type IdEntity struct {
-	ID        uint32    `gorm:"primaryKey"`
+	ID        int64     `gorm:"primaryKey"`
 	CreatedAt time.Time `gorm:"autoCreateTime;not null"`
 	UpdatedAt time.Time `gorm:"autoUpdateTime;not null"`
 }
